chore: extract getSwapperBySwapperName helper
