+++ src/components/MultiHopTrade/helpers.ts
@@ -1,4 +1,19 @@
 import { getMaybeCompositeAssetSymbol } from 'lib/mixpanel/helpers'
+import type { Swapper2, Swapper2Api } from 'lib/swapper/api'
+import { SwapperName } from 'lib/swapper/api'
+import { cowSwapper } from 'lib/swapper/swappers/CowSwapper/CowSwapper2'
+import { cowApi } from 'lib/swapper/swappers/CowSwapper/endpoints'
+import { lifiApi } from 'lib/swapper/swappers/LifiSwapper/endpoints'
+import { lifiSwapper } from 'lib/swapper/swappers/LifiSwapper/LifiSwapper2'
+import { oneInchApi } from 'lib/swapper/swappers/OneInchSwapper/endpoints'
+import { oneInchSwapper } from 'lib/swapper/swappers/OneInchSwapper/OneInchSwapper2'
+import { osmosisApi } from 'lib/swapper/swappers/OsmosisSwapper/endpoints'
+import { osmosisSwapper } from 'lib/swapper/swappers/OsmosisSwapper/OsmosisSwapper2'
+import { thorchainApi } from 'lib/swapper/swappers/ThorchainSwapper/endpoints'
+import { thorchainSwapper } from 'lib/swapper/swappers/ThorchainSwapper/ThorchainSwapper2'
+import { zrxApi } from 'lib/swapper/swappers/ZrxSwapper/endpoints'
+import { zrxSwapper } from 'lib/swapper/swappers/ZrxSwapper/ZrxSwapper2'
+import { assertUnreachable } from 'lib/utils'
 import type { ReduxState } from 'state/reducer'
 import { selectAssets, selectWillDonate } from 'state/slices/selectors'
 import {
@@ -48,5 +63,26 @@ export const getMixpanelEventData = () => {
     donationAmountUserCurrency,
     [compositeBuyAsset]: buyAmountBeforeFeesCryptoPrecision,
     [compositeSellAsset]: sellAmountBeforeFeesCryptoPrecision,
+  }
+}
+
+export const getSwapperBySwapperName = (swapperName: SwapperName): Swapper2 & Swapper2Api => {
+  switch (swapperName) {
+    case SwapperName.LIFI:
+      return { ...lifiSwapper, ...lifiApi }
+    case SwapperName.Thorchain:
+      return { ...thorchainSwapper, ...thorchainApi }
+    case SwapperName.Zrx:
+      return { ...zrxSwapper, ...zrxApi }
+    case SwapperName.CowSwap:
+      return { ...cowSwapper, ...cowApi }
+    case SwapperName.OneInch:
+      return { ...oneInchSwapper, ...oneInchApi }
+    case SwapperName.Osmosis:
+      return { ...osmosisSwapper, ...osmosisApi }
+    case SwapperName.Test:
+      throw Error('not implemented')
+    default:
+      assertUnreachable(swapperName)
   }
 }
